Improves resource and settlement placement
Improves the placement of resource and settlement icons and
labels on the map, especially when both are present on the same tile.

It also fixes an issue where multiple settlement names could appear
on a tile by only rendering the first one and flagging it if there
are multiple names.

Updates scouted label offset and settlement label y offset.

Increments version to 0.56.0.